feat: pagination patient

+++ HealthCare/Services/PatientService.cs
@@ -1,11 +1,8 @@
 ﻿using AutoMapper;
 using HealthCare.Infrastructure.Configurations.Jwt;
 using HealthCare.Infrastructure.Managers.Interfaces;
-using HealthCare.Models.AddressEntity;
-using HealthCare.Models.AddressEntity.Dto;
 using HealthCare.Models.PatientEntity;
 using HealthCare.Models.PatientEntity.Dto;
-using HealthCare.Models.ProfessionalPatientEntity;
 using HealthCare.Repositories.Interfaces;
 using HealthCare.Services.Interfaces;
 using HealthCare.Utils;
@@ -26,7 +23,6 @@ public class PatientService(
     public async Task<PatientResponseDto> Create(PatientRequestDto newPatient)
     {
         var user = await accessor.GetEmployee(jwtBody, tokenService, managerUow);
-
         user = await managerUow.UserManager.FindByIdAsync(Convert.ToString(user.Id) ?? string.Empty) ??
                throw new BadHttpRequestException("User not found.");
 
@@ -44,16 +40,12 @@ public class PatientService(
         return mapper.Map<PatientResponseDto>(await repositoryUow.PatientRepository.Create(patient));
     }
 
-    public async Task<PatientResponseDto> Update(PatientRequestDto patientRequestDto)
+    public async Task<PatientResponseDto> Update(Guid patientId, PatientUpdateDto updatePatient)
     {
-        var patient = await repositoryUow.PatientRepository.FindByCpf(patientRequestDto.Cpf) ??
-                      throw new BadHttpRequestException("Patient not found.");
-
-        patient.Name = patientRequestDto.Name;
-        patient.DateOfBirth = patientRequestDto.DateOfBirth;
-        patient.Address = addressService.Update(patientRequestDto.Address);
-        patient.Nationality = patientRequestDto.Nationality;
-        patient.MaritalStatus = patientRequestDto.MaritalStatus;
+        var patient = await repositoryUow.PatientRepository.GetPatient(patientId);
+        if (patient == null) throw new BadHttpRequestException("Patient not found.");
+        patient = mapper.Map(updatePatient, patient);
+        patient.Address = await addressService.Update(patient.AddressId, updatePatient.Address);
 
         repositoryUow.PatientRepository.Update(patient);
         await repositoryUow.CommitAsync();
@@ -61,17 +53,35 @@ public class PatientService(
         return mapper.Map<PatientResponseDto>(await repositoryUow.PatientRepository.Create(patient));
     }
 
-    public async Task<PatientResponseDto> GetPatient(string id)
+    public async Task<PatientResponseDto> GetPatient(Guid id)
     {
         return mapper.Map<PatientResponseDto>(await repositoryUow.PatientRepository.GetPatient(id) ??
                                               throw new BadHttpRequestException("Patient not found."));
     }
 
-    public async Task Delete(string cpf)
+    public async Task Delete(Guid patientId)
     {
-        var patient = await repositoryUow.PatientRepository.FindByCpf(cpf) ??
-                      throw new BadHttpRequestException("Patient not found.");
+        var patient = await repositoryUow.PatientRepository.GetPatient(patientId);
+        if (patient == null) throw new BadHttpRequestException("Patient not found.");
+
+        await addressService.Delete(patient.AddressId);
         repositoryUow.PatientRepository.Delete(patient);
+
         await repositoryUow.CommitAsync();
+    }
+
+    public async Task<PatientPageResult> Patients(int page, int pageSize)
+    {
+        var user = await accessor.GetEmployee(jwtBody, tokenService, managerUow);
+        user = await managerUow.UserManager.FindByIdAsync(Convert.ToString(user.Id) ?? string.Empty) ??
+               throw new BadHttpRequestException("User not found.");
+
+        var patients = await repositoryUow.PatientRepository.GetPatients(user.Id, page, pageSize);
+
+        return new()
+        {
+            Patients = mapper.Map<List<PatientResponseDto>>(patients),
+            TotalCount = patients.Count
+        };
     }
 }
